requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.39.1
Django==5.0.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.7
inflection==0.5.1
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0